Consolidate ActivityNet Captions retrieval tasks into a single file
Made-with: Cursor
